move common string
